+++ partiql-spi/src/main/java/org/partiql/spi/value/DatumBytes.java
@@ -3,6 +3,8 @@ package org.partiql.spi.value;
 import org.jetbrains.annotations.NotNull;
 import org.partiql.spi.types.PType;
 
+import java.util.StringJoiner;
+
 /**
  * This shall always be package-private (internal).
  * <p></p>
@@ -33,5 +35,19 @@ class DatumBytes implements Datum {
     @Override
     public PType getType() {
         return _type;
+    }
+
+    @Override
+    public String toString() {
+        StringJoiner joiner = new StringJoiner(", ", "[", "]");
+        for (byte b : _value) {
+            joiner.add(Byte.toString(b));
+        }
+
+        return "DatumBytes{" +
+                "_type=" + _type +
+                ", _value=" + joiner +
+                '}';
+
     }
 }